Add files via upload
This repository contains all the functions and scripts used in the Orbital Mechanics Project

The project was divided in two assigments. Assigment 1 is about the interplanetary mission, Assigment 2 instead was about the planetary mission. The main scrips are "PlanetaryMission_group_2157.m" and "InterplanetaryMission_group_2157.m". The remaining scripts are functions, already divided for their usage in the first or second assigment.